+++ gradle-extensions-subprojects/gradle-properties-extensions/main/kotlin/com/javiersc/gradle/properties/extensions/logger.kt
@@ -1,0 +1,10 @@
+package com.javiersc.gradle.properties.extensions
+
+import org.gradle.api.logging.Logger
+import org.gradle.api.logging.Logging
+
+internal val defaultLogger: Logger = Logging.getLogger("Gradle Property Extensions")
+
+internal fun propertyNotFound(name: String, source: String) {
+    defaultLogger.debug("Property $name found in `$source`")
+}